requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
ci-info==0.3.0
click==8.1.3
colorama==0.4.5
comm==0.1.2
configobj==5.0.8
configparser==5.3.0
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.4
etelemetry==0.3.0
executing==1.2.0
fastapi==0.94.0
fastjsonschema==2.16.2
filelock==3.9.0
fonttools==4.37.1
fqdn==1.5.1
frozenlist==1.4.0
future==0.18.3
greenlet==3.0.0
h11==0.13.0
httplib2==0.21.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.20.1
ipython==8.8.0
ipython-genutils==0.2.0
isodate==0.6.1
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jotform==1.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.18.0
kiwisolver==1.4.4
loguru==0.6.0
looseversion==1.1.2
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nibabel==5.0.1
nipype==1.8.5
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.3
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pendulum==2.1.2
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.11
platformdirs==2.6.2
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
prov==2.0.0
psutil==5.9.4
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
pydantic==1.9.1
pydot==1.4.2
Pygments==2.14.0
pyhumps==3.8.0
PyMuPDF==1.21.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.1.3
pytest-asyncio==0.21.1
pytest-dotenv==0.5.2
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.4
pytz==2022.2.1
pytzdata==2020.1
pywin32==305
pywinpty==2.0.10
pyxnat==1.5
PyYAML==6.0
pyzmq==25.0.0
rdflib==6.2.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scipy==1.9.3
seaborn==0.12.0
Send2Trash==1.8.0
setuptools-scm==6.4.2
simplejson==3.18.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
stack-data==0.6.2
starlette==0.26.0.post1
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.8.1
traits==6.3.2
typing_extensions==4.3.0
uri-template==1.2.0
urllib3==1.26.14
uvicorn==0.18.2
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
win32-setctime==1.1.0
xlrd==2.0.1
yarg==0.1.9
yarl==1.9.2
